+++ lib/whois/definitions/tlds.rb
@@ -60,29 +60,29 @@ Whois::Server.define :tld, ".am", "whois.nic.am"
 Whois::Server.define :tld, ".an", nil, { :adapter => :none }
 Whois::Server.define :tld, ".ao", nil, { :adapter => :none }
 Whois::Server.define :tld, ".aq", nil, { :adapter => :none }
-Whois::Server.define :tld, ".ar", nil, { :web => "http://www.nic.ar/", :adapter => :web }
+Whois::Server.define :tld, ".ar", nil, { :adapter => :web, :url => "http://www.nic.ar/" }
 Whois::Server.define :tld, ".as", "whois.nic.as"
 Whois::Server.define :tld, ".priv.at", "whois.nic.priv.at"
 Whois::Server.define :tld, ".at", "whois.nic.at"
 Whois::Server.define :tld, ".au", "whois.audns.net.au"
 Whois::Server.define :tld, ".aw", nil, { :adapter => :none }
 Whois::Server.define :tld, ".ax", nil, { :adapter => :none }
-Whois::Server.define :tld, ".az", nil, { :web => "http://www.nic.az/AzCheck.htm", :adapter => :web }
-Whois::Server.define :tld, ".ba", nil, { :web => "http://www.nic.ba/stream/whois/", :adapter => :web }
-Whois::Server.define :tld, ".bb", nil, { :web => "http://whois.telecoms.gov.bb/search_domain.php", :adapter => :web }
-Whois::Server.define :tld, ".bd", nil, { :adapter => :web, :web => "http://whois.btcl.net.bd/"}
+Whois::Server.define :tld, ".az", nil, { :adapter => :web, :url => "http://www.nic.az/AzCheck.htm" }
+Whois::Server.define :tld, ".ba", nil, { :adapter => :web, :url => "http://www.nic.ba/stream/whois/" }
+Whois::Server.define :tld, ".bb", nil, { :adapter => :web, :url => "http://whois.telecoms.gov.bb/search_domain.php" }
+Whois::Server.define :tld, ".bd", nil, { :adapter => :web, :url => "http://whois.btcl.net.bd/"}
 Whois::Server.define :tld, ".be", "whois.dns.be"
 Whois::Server.define :tld, ".bf", nil, { :adapter => :none }
 Whois::Server.define :tld, ".bg", "whois.register.bg"
 Whois::Server.define :tld, ".bh", nil, { :adapter => :none }
-Whois::Server.define :tld, ".bi", nil, { :web => "https://whois.nic.bi/whois.php", :adapter => :web }
+Whois::Server.define :tld, ".bi", nil, { :adapter => :web, :url => "https://whois.nic.bi/whois.php" }
 Whois::Server.define :tld, ".bj", "whois.nic.bj"
-Whois::Server.define :tld, ".bm", nil, { :web => "http://207.228.133.14/cgi-bin/lansaweb?procfun+BMWHO+BMWHO2+WHO", :adapter => :web }
+Whois::Server.define :tld, ".bm", nil, { :adapter => :web, :url => "http://207.228.133.14/cgi-bin/lansaweb?procfun+BMWHO+BMWHO2+WHO" }
 Whois::Server.define :tld, ".bn", nil, { :adapter => :none }
 Whois::Server.define :tld, ".bo", "whois.nic.bo"
 Whois::Server.define :tld, ".br", "whois.registro.br"
-Whois::Server.define :tld, ".bs", nil, { :web => "http://www.nic.bs/cgi-bin/search.pl", :adapter => :web }
-Whois::Server.define :tld, ".bt", nil, { :web => "http://www.nic.bt/", :adapter => :web }
+Whois::Server.define :tld, ".bs", nil, { :adapter => :web, :url => "http://www.nic.bs/cgi-bin/search.pl" }
+Whois::Server.define :tld, ".bt", nil, { :adapter => :web, :url => "http://www.nic.bt/" }
 Whois::Server.define :tld, ".bv", nil, { :adapter => :none }
 Whois::Server.define :tld, ".by", 'whois.cctld.by'
 Whois::Server.define :tld, ".bw", nil, { :adapter => :none }
@@ -92,7 +92,7 @@ Whois::Server.define :tld, ".ca", "whois.cira.ca"
 Whois::Server.define :tld, ".cc", "whois.nic.cc", { :adapter => :verisign }
 Whois::Server.define :tld, ".cd", "whois.nic.cd"
 Whois::Server.define :tld, ".cf", nil, { :adapter => :none }
-Whois::Server.define :tld, ".cg", nil, { :web => "http://www.nic.cg/cgi-bin/whois.pl", :adapter => :web }
+Whois::Server.define :tld, ".cg", nil, { :adapter => :web, :url => "http://www.nic.cg/cgi-bin/whois.pl" }
 Whois::Server.define :tld, ".ch", "whois.nic.ch"
 Whois::Server.define :tld, ".ci", "whois.nic.ci"
 Whois::Server.define :tld, ".ck", "whois.nic.ck"
@@ -101,29 +101,29 @@ Whois::Server.define :tld, ".cm", "whois.netcom.cm"
 Whois::Server.define :tld, ".edu.cn", "whois.edu.cn"
 Whois::Server.define :tld, ".cn", "whois.cnnic.cn"
 Whois::Server.define :tld, ".co", "whois.nic.co"
-Whois::Server.define :tld, ".cr", nil, { :web => "http://www.nic.cr/niccr_publico/showRegistroDominiosScreen.do", :adapter => :web }
-Whois::Server.define :tld, ".cu", nil, { :adapter => :web, :web => "http://www.nic.cu/" }
-Whois::Server.define :tld, ".cv", nil, { :adapter => :web, :web => "http://www.dns.cv/" }
+Whois::Server.define :tld, ".cr", nil, { :adapter => :web, :url => "http://www.nic.cr/niccr_publico/showRegistroDominiosScreen.do" }
+Whois::Server.define :tld, ".cu", nil, { :adapter => :web, :url => "http://www.nic.cu/" }
+Whois::Server.define :tld, ".cv", nil, { :adapter => :web, :url => "http://www.dns.cv/" }
 Whois::Server.define :tld, ".cx", "whois.nic.cx"
-Whois::Server.define :tld, ".cy", nil, { :web => "http://www.nic.cy/nslookup/online_database.php", :adapter => :web }
+Whois::Server.define :tld, ".cy", nil, { :adapter => :web, :url => "http://www.nic.cy/nslookup/online_database.php" }
 Whois::Server.define :tld, ".cz", "whois.nic.cz"
 Whois::Server.define :tld, ".de", "whois.denic.de", { :adapter => :formatted, :format => "-T dn,ace %s" }
-Whois::Server.define :tld, ".dj", nil, { :adapter => :web, :web => "http://www.nic.dj/whois.php" }
+Whois::Server.define :tld, ".dj", nil, { :adapter => :web, :url => "http://www.nic.dj/whois.php" }
 Whois::Server.define :tld, ".dk", "whois.dk-hostmaster.dk", { :adapter => :formatted, :format => "--show-handles %s" }
 Whois::Server.define :tld, ".dm", "whois.nic.dm"
-Whois::Server.define :tld, ".do", nil, { :web => "http://www.nic.do/whois-h.php3", :adapter => :web }
+Whois::Server.define :tld, ".do", nil, { :adapter => :web, :url => "http://www.nic.do/whois-h.php3" }
 Whois::Server.define :tld, ".dz", "whois.nic.dz"
 Whois::Server.define :tld, ".ec", "whois.nic.ec"
 Whois::Server.define :tld, ".ee", "whois.tld.ee"
-Whois::Server.define :tld, ".eg", nil, { :web => "http://lookup.egregistry.eg/english.aspx", :adapter => :web }
+Whois::Server.define :tld, ".eg", nil, { :adapter => :web, :url => "http://lookup.egregistry.eg/english.aspx" }
 Whois::Server.define :tld, ".er", nil, { :adapter => :none }
-Whois::Server.define :tld, ".es", nil, { :web => "https://www.nic.es/", :adapter => :web }
+Whois::Server.define :tld, ".es", nil, { :adapter => :web, :url => "https://www.nic.es/" }
 Whois::Server.define :tld, ".et", nil, { :adapter => :none }
 Whois::Server.define :tld, ".eu", "whois.eu"
 Whois::Server.define :tld, ".fi", "whois.fi"
 Whois::Server.define :tld, ".fj", "whois.usp.ac.fj"
 Whois::Server.define :tld, ".fk", nil, { :adapter => :none }
-Whois::Server.define :tld, ".fm", nil, { :web => "http://www.dot.fm/whois.html", :adapter => :web }
+Whois::Server.define :tld, ".fm", nil, { :adapter => :web, :url => "http://www.dot.fm/whois.html" }
 Whois::Server.define :tld, ".fo", "whois.ripe.net"
 Whois::Server.define :tld, ".aeroport.fr", "whois.smallregistry.net"
 Whois::Server.define :tld, ".avocat.fr", "whois.smallregistry.net"
@@ -140,20 +140,20 @@ Whois::Server.define :tld, ".fr", "whois.nic.fr"
 Whois::Server.define :tld, ".ga", nil, { :adapter => :none }
 Whois::Server.define :tld, ".gb", nil, { :adapter => :none }
 Whois::Server.define :tld, ".gd", "whois.adamsnames.tc"
-Whois::Server.define :tld, ".ge", nil, { :adapter => :web, :web => "http://www.registration.ge/" }
-Whois::Server.define :tld, ".gf", nil, { :adapter => :web, :web => "https://www.dom-enic.com/whois.html" }
+Whois::Server.define :tld, ".ge", nil, { :adapter => :web, :url => "http://www.registration.ge/" }
+Whois::Server.define :tld, ".gf", nil, { :adapter => :web, :url => "https://www.dom-enic.com/whois.html" }
 Whois::Server.define :tld, ".gg", "whois.gg"
-Whois::Server.define :tld, ".gh", nil, { :adapter => :web, :web => "http://www.nic.gh/customer/search_c.htm" }
+Whois::Server.define :tld, ".gh", nil, { :adapter => :web, :url => "http://www.nic.gh/customer/search_c.htm" }
 Whois::Server.define :tld, ".gi", "whois.afilias-grs.info", { :adapter => :afilias }
 Whois::Server.define :tld, ".gl", "whois.nic.gl"
-Whois::Server.define :tld, ".gm", nil, { :adapter => :web, :web => "http://www.nic.gm/htmlpages/whois.htm" }
+Whois::Server.define :tld, ".gm", nil, { :adapter => :web, :url => "http://www.nic.gm/htmlpages/whois.htm" }
 Whois::Server.define :tld, ".gn", nil, { :adapter => :none }
 Whois::Server.define :tld, ".gp", nil, { :adapter => :none }
 Whois::Server.define :tld, ".gq", nil, { :adapter => :none }
-Whois::Server.define :tld, ".gr", nil, { :adapter => :web, :web => "https://grweb.ics.forth.gr/whois_en.jsp" }
+Whois::Server.define :tld, ".gr", nil, { :adapter => :web, :url => "https://grweb.ics.forth.gr/whois_en.jsp" }
 Whois::Server.define :tld, ".gs", "whois.nic.gs"
-Whois::Server.define :tld, ".gt", nil, { :adapter => :web, :web => "http://www.gt/whois.html" }
-Whois::Server.define :tld, ".gu", nil, { :adapter => :web, :web => "http://gadao.gov.gu/domainsearch.htm" }
+Whois::Server.define :tld, ".gt", nil, { :adapter => :web, :url => "http://www.gt/whois.html" }
+Whois::Server.define :tld, ".gu", nil, { :adapter => :web, :url => "http://gadao.gov.gu/domainsearch.htm" }
 Whois::Server.define :tld, ".gw", nil, { :adapter => :none }
 Whois::Server.define :tld, ".gy", "whois.registry.gy"
 Whois::Server.define :tld, ".hk", "whois.hkirc.hk"
@@ -174,26 +174,26 @@ Whois::Server.define :tld, ".is", "whois.isnic.is"
 Whois::Server.define :tld, ".it", "whois.nic.it"
 Whois::Server.define :tld, ".je", "whois.je"
 Whois::Server.define :tld, ".jm", nil, { :adapter => :none }
-Whois::Server.define :tld, ".jo", nil, { :web => "http://www.dns.jo/Whois.aspx", :adapter => :web }
+Whois::Server.define :tld, ".jo", nil, { :adapter => :web, :url => "http://www.dns.jo/Whois.aspx" }
 Whois::Server.define :tld, ".jp", "whois.jprs.jp", { :adapter => :formatted, :format => "%s/e" }
 Whois::Server.define :tld, ".ke", "whois.kenic.or.ke"
 Whois::Server.define :tld, ".kg", "whois.domain.kg"
 Whois::Server.define :tld, ".kh", nil, { :adapter => :none }
 Whois::Server.define :tld, ".ki", "whois.nic.ki"
 Whois::Server.define :tld, ".km", nil, { :adapter => :none }
-Whois::Server.define :tld, ".kn", nil, { :web => "http://www.nic.kn/", :adapter => :web }
+Whois::Server.define :tld, ".kn", nil, { :adapter => :web, :url => "http://www.nic.kn/" }
 Whois::Server.define :tld, ".kp", nil, { :adapter => :none }
 Whois::Server.define :tld, ".kr", "whois.nic.or.kr"
-Whois::Server.define :tld, ".kw", nil, { :web => "http://www.kw/", :adapter => :web }
-Whois::Server.define :tld, ".ky", nil, { :web => "http://kynseweb.messagesecure.com/kywebadmin/", :adapter => :web }
+Whois::Server.define :tld, ".kw", nil, { :adapter => :web, :url => "http://www.kw/" }
+Whois::Server.define :tld, ".ky", nil, { :adapter => :web, :url => "http://kynseweb.messagesecure.com/kywebadmin/" }
 Whois::Server.define :tld, ".kz", "whois.nic.kz"
 Whois::Server.define :tld, ".la", "whois.nic.la"
-Whois::Server.define :tld, ".lb", nil, { :adapter => :web, :web => "http://www.aub.edu.lb/lbdr/"}
+Whois::Server.define :tld, ".lb", nil, { :adapter => :web, :url => "http://www.aub.edu.lb/lbdr/"}
 Whois::Server.define :tld, ".lc", "whois.afilias-grs.info", { :adapter => :afilias }
 Whois::Server.define :tld, ".li", "whois.nic.li"
 Whois::Server.define :tld, ".lk", "whois.nic.lk"
 Whois::Server.define :tld, ".lr", nil, { :adapter => :none }
-Whois::Server.define :tld, ".ls", nil, { :adapter => :web, :web => "http://www.co.ls/co.asp"}
+Whois::Server.define :tld, ".ls", nil, { :adapter => :web, :url => "http://www.co.ls/co.asp"}
 Whois::Server.define :tld, ".lt", "whois.domreg.lt"
 Whois::Server.define :tld, ".lu", "whois.dns.lu"
 Whois::Server.define :tld, ".lv", "whois.nic.lv"
@@ -204,19 +204,19 @@ Whois::Server.define :tld, ".md", "whois.nic.md"
 Whois::Server.define :tld, ".me", "whois.meregistry.net"
 Whois::Server.define :tld, ".mg", "whois.nic.mg"
 Whois::Server.define :tld, ".mh", nil, { :adapter => :none }
-Whois::Server.define :tld, ".mk", nil, { :adapter => :web, :web => "http://dns.marnet.net.mk/registar.php" }
+Whois::Server.define :tld, ".mk", nil, { :adapter => :web, :url => "http://dns.marnet.net.mk/registar.php" }
 Whois::Server.define :tld, ".ml", nil, { :adapter => :none }
 Whois::Server.define :tld, ".mm", nil, { :adapter => :none }
 Whois::Server.define :tld, ".mn", "whois.afilias-grs.info", { :adapter => :afilias }
-Whois::Server.define :tld, ".mo", nil, { :web => "http://www.monic.net.mo/", :adapter => :web }
+Whois::Server.define :tld, ".mo", nil, { :adapter => :web, :url => "http://www.monic.net.mo/" }
 Whois::Server.define :tld, ".mp", nil, { :adapter => :none }
-Whois::Server.define :tld, ".mq", nil, { :adapter => :web, :web => "https://www.dom-enic.com/whois.html" }
+Whois::Server.define :tld, ".mq", nil, { :adapter => :web, :url => "https://www.dom-enic.com/whois.html" }
 Whois::Server.define :tld, ".mr", nil, { :adapter => :none }
 Whois::Server.define :tld, ".ms", "whois.nic.ms"
-Whois::Server.define :tld, ".mt", nil, { :web => "https://www.nic.org.mt/dotmt/", :adapter => :web }
+Whois::Server.define :tld, ".mt", nil, { :adapter => :web, :url => "https://www.nic.org.mt/dotmt/" }
 Whois::Server.define :tld, ".mu", "whois.nic.mu"
 Whois::Server.define :tld, ".mv", nil, { :adapter => :none }
-Whois::Server.define :tld, ".mw", nil, { :web => "http://www.registrar.mw/", :adapter => :web }
+Whois::Server.define :tld, ".mw", nil, { :adapter => :web, :url => "http://www.registrar.mw/" }
 Whois::Server.define :tld, ".mx", "whois.nic.mx"
 Whois::Server.define :tld, ".my", "whois.domainregistry.my"
 Whois::Server.define :tld, ".mz", nil, { :adapter => :none }
@@ -225,36 +225,36 @@ Whois::Server.define :tld, ".nc", "whois.nc"
 Whois::Server.define :tld, ".ne", nil, { :adapter => :none }
 Whois::Server.define :tld, ".nf", "whois.nic.net.nf"
 Whois::Server.define :tld, ".ng", "whois.nic.net.ng"
-Whois::Server.define :tld, ".ni", nil, { :adapter => :web, :web => "http://www.nic.ni/"}
+Whois::Server.define :tld, ".ni", nil, { :adapter => :web, :url => "http://www.nic.ni/"}
 Whois::Server.define :tld, ".nl", "whois.domain-registry.nl"
 Whois::Server.define :tld, ".no", "whois.norid.no"
-Whois::Server.define :tld, ".np", nil, { :web => "http://register.mos.com.np/userSearchInc.asp", :adapter => :web }
-Whois::Server.define :tld, ".nr", nil, { :web => "http://www.cenpac.net.nr/dns/whois.html", :adapter => :web }
+Whois::Server.define :tld, ".np", nil, { :adapter => :web, :url => "http://register.mos.com.np/userSearchInc.asp" }
+Whois::Server.define :tld, ".nr", nil, { :adapter => :web, :url => "http://www.cenpac.net.nr/dns/whois.html" }
 Whois::Server.define :tld, ".nu", "whois.nic.nu"
 Whois::Server.define :tld, ".nz", "whois.srs.net.nz"
 Whois::Server.define :tld, ".om", "whois.registry.om"
-Whois::Server.define :tld, ".pa", nil, { :web => "http://www.nic.pa/", :adapter => :web }
+Whois::Server.define :tld, ".pa", nil, { :adapter => :web, :url => "http://www.nic.pa/" }
 Whois::Server.define :tld, ".pe", "kero.yachay.pe"
 Whois::Server.define :tld, ".pf", nil, { :adapter => :none }
 Whois::Server.define :tld, ".pg", nil, { :adapter => :none }
-Whois::Server.define :tld, ".ph", nil, { :web => "http://www.dot.ph/", :adapter => :web }
-Whois::Server.define :tld, ".pk", nil, { :web => "http://www.pknic.net.pk/", :adapter => :web }
+Whois::Server.define :tld, ".ph", nil, { :adapter => :web, :url => "http://www.dot.ph/" }
+Whois::Server.define :tld, ".pk", nil, { :adapter => :web, :url => "http://www.pknic.net.pk/" }
 Whois::Server.define :tld, ".co.pl", "whois.co.pl"
 Whois::Server.define :tld, ".pl", "whois.dns.pl"
 Whois::Server.define :tld, ".pm", "whois.nic.fr"
-Whois::Server.define :tld, ".pn", nil, { :web => "http://www.pitcairn.pn/PnRegistry/", :adapter => :web }
+Whois::Server.define :tld, ".pn", nil, { :adapter => :web, :url => "http://www.pitcairn.pn/PnRegistry/" }
 Whois::Server.define :tld, ".pr", "whois.nic.pr"
 Whois::Server.define :tld, ".ps", "whois.pnina.ps"
 Whois::Server.define :tld, ".pt", "whois.dns.pt"
 Whois::Server.define :tld, ".pw", nil, { :adapter => :none }
-Whois::Server.define :tld, ".py", nil, { :web => "http://www.nic.py/consultas.html", :adapter => :web }
+Whois::Server.define :tld, ".py", nil, { :adapter => :web, :url => "http://www.nic.py/consultas.html" }
 Whois::Server.define :tld, ".qa", "whois.registry.qa"
 Whois::Server.define :tld, ".re", "whois.nic.fr"
 Whois::Server.define :tld, ".ro", "whois.rotld.ro"
 Whois::Server.define :tld, ".rs", "whois.rnids.rs"
 Whois::Server.define :tld, ".edu.ru", "whois.informika.ru"
 Whois::Server.define :tld, ".ru", "whois.tcinet.ru"
-Whois::Server.define :tld, ".rw", nil, { :adapter => :web, :web => "http://www.nic.rw/cgi-bin/whois.pl"}
+Whois::Server.define :tld, ".rw", nil, { :adapter => :web, :url => "http://www.nic.rw/cgi-bin/whois.pl"}
 Whois::Server.define :tld, ".sa", "saudinic.net.sa"
 Whois::Server.define :tld, ".sb", "whois.nic.net.sb"
 Whois::Server.define :tld, ".sc", "whois.afilias-grs.info", { :adapter => :afilias }
@@ -272,16 +272,16 @@ Whois::Server.define :tld, ".so", "whois.nic.so"
 Whois::Server.define :tld, ".sr", nil, { :adapter => :none }
 Whois::Server.define :tld, ".st", "whois.nic.st"
 Whois::Server.define :tld, ".su", "whois.tcinet.ru"
-Whois::Server.define :tld, ".sv", nil, { :adapter => :web, :web => "http://www.uca.edu.sv/dns/" }
+Whois::Server.define :tld, ".sv", nil, { :adapter => :web, :url => "http://www.uca.edu.sv/dns/" }
 Whois::Server.define :tld, ".sx", "whois.sx"
 Whois::Server.define :tld, ".sy", nil, { :adapter => :none }
 Whois::Server.define :tld, ".sz", nil, { :adapter => :none }
 Whois::Server.define :tld, ".tc", "whois.adamsnames.tc"
 Whois::Server.define :tld, ".td", nil, { :adapter => :none }
 Whois::Server.define :tld, ".tf", "whois.nic.fr"
-Whois::Server.define :tld, ".tg", nil, { :web => "http://www.nic.tg/", :adapter => :web }
+Whois::Server.define :tld, ".tg", nil, { :adapter => :web, :url => "http://www.nic.tg/" }
 Whois::Server.define :tld, ".th", "whois.thnic.co.th"
-Whois::Server.define :tld, ".tj", nil, { :web => "http://www.nic.tj/whois.html", :adapter => :web }
+Whois::Server.define :tld, ".tj", nil, { :adapter => :web, :url => "http://www.nic.tj/whois.html" }
 Whois::Server.define :tld, ".tk", "whois.dot.tk"
 Whois::Server.define :tld, ".tl", "whois.nic.tl"
 Whois::Server.define :tld, ".tm", "whois.nic.tm"
@@ -289,7 +289,7 @@ Whois::Server.define :tld, ".tn", "whois.ati.tn"
 Whois::Server.define :tld, ".to", "whois.tonic.to"
 Whois::Server.define :tld, ".tp", nil, { :adapter => :none }
 Whois::Server.define :tld, ".tr", "whois.nic.tr"
-Whois::Server.define :tld, ".tt", nil, { :web => "http://www.nic.tt/cgi-bin/search.pl", :adapter => :web }
+Whois::Server.define :tld, ".tt", nil, { :adapter => :web, :url => "http://www.nic.tt/cgi-bin/search.pl" }
 Whois::Server.define :tld, ".tv", "whois.nic.tv", { :adapter => :verisign }
 Whois::Server.define :tld, ".tw", "whois.twnic.net.tw"
 Whois::Server.define :tld, ".tz", "whois.tznic.or.tz"
@@ -309,16 +309,16 @@ Whois::Server.define :tld, ".parliament.uk", nil, { :adapter => :none }
 Whois::Server.define :tld, ".police.uk", nil, { :adapter => :none }
 Whois::Server.define :tld, ".uk", "whois.nic.uk"
 Whois::Server.define :tld, ".us", "whois.nic.us"
-Whois::Server.define :tld, ".com.uy", nil, { :web => "https://nic.anteldata.com.uy/dns/", :adapter => :web }
+Whois::Server.define :tld, ".com.uy", nil, { :adapter => :web, :url => "https://nic.anteldata.com.uy/dns/" }
 Whois::Server.define :tld, ".uy", "whois.nic.org.uy"
 Whois::Server.define :tld, ".uz", "whois.cctld.uz"
 Whois::Server.define :tld, ".va", nil, { :adapter => :none }
 Whois::Server.define :tld, ".vc", "whois.afilias-grs.info", { :adapter => :afilias }
 Whois::Server.define :tld, ".ve", "whois.nic.ve"
 Whois::Server.define :tld, ".vg", "whois.adamsnames.tc"
-Whois::Server.define :tld, ".vi", nil, { :web => "http://www.nic.vi/whoisform.htm", :adapter => :web }
-Whois::Server.define :tld, ".vn", nil, { :web => "http://www.vnnic.vn/english/", :adapter => :web }
-Whois::Server.define :tld, ".vu", nil, { :web => "http://www.vunic.vu/whois.html", :adapter => :web }
+Whois::Server.define :tld, ".vi", nil, { :adapter => :web, :url => "http://www.nic.vi/whoisform.htm" }
+Whois::Server.define :tld, ".vn", nil, { :adapter => :web, :url => "http://www.vnnic.vn/english/" }
+Whois::Server.define :tld, ".vu", nil, { :adapter => :web, :url => "http://www.vunic.vu/whois.html" }
 Whois::Server.define :tld, ".wf", "whois.nic.fr"
 Whois::Server.define :tld, ".ws", "whois.samoanic.ws"
 Whois::Server.define :tld, ".ye", nil, { :adapter => :none }
@@ -347,7 +347,7 @@ Whois::Server.define :tld, ".xn--kprw13d", "whois.twnic.net.tw"
 Whois::Server.define :tld, ".xn--kpry57d", "whois.twnic.net.tw"
 Whois::Server.define :tld, ".xn--lgbbat1ad8j", "whois.nic.dz"
 Whois::Server.define :tld, ".xn--mgbaam7a8h", "whois.aeda.net.ae"
-Whois::Server.define :tld, ".xn--mgbayh7gpa", nil, { :adapter => :web, :web => "http://idn.jo/whois_a.aspx" }
+Whois::Server.define :tld, ".xn--mgbayh7gpa", nil, { :adapter => :web, :url => "http://idn.jo/whois_a.aspx" }
 Whois::Server.define :tld, ".xn--mgbbh1a71e", nil, { :adapter => :none }
 Whois::Server.define :tld, ".xn--mgbc0a9azcg", nil, { :adapter => :none }
 Whois::Server.define :tld, ".xn--mgberp4a5d4ar", "whois.nic.net.sa"