+++ spec/whois/server/adapters/web_spec.rb
@@ -3,7 +3,7 @@ require "spec_helper"
 describe Whois::Server::Adapters::Web do
 
   before(:each) do
-    @definition = [:tld, ".test", nil, { :web => "http://whois.test" }]
+    @definition = [:tld, ".test", nil, { :url => "http://whois.test" }]
   end